CLI biosimulations command (is default command from docker)
