+++ vcell-cli/src/main/java/org/vcell/cli/CLIStandalone.java
@@ -1,5 +1,6 @@
 package org.vcell.cli;
 
+import org.vcell.cli.biosimulation.BiosimulationsCommand;
 import org.vcell.cli.run.ExecuteCommand;
 import org.vcell.cli.vcml.ConvertCommand;
 
@@ -8,7 +9,13 @@ import picocli.CommandLine;
 import picocli.CommandLine.Command;
 
 
-@Command(name = "CLIStandalone", subcommands = {ConvertCommand.class, ExecuteCommand.class, VersionCommand.class, CommandLine.HelpCommand.class})
+@Command(name = "CLIStandalone", subcommands = {
+        BiosimulationsCommand.class,
+        ConvertCommand.class,
+        ExecuteCommand.class,
+        VersionCommand.class,
+        CommandLine.HelpCommand.class
+})
 public class CLIStandalone {
 
     public static void main(String[] args) {
